Renamed app.js to postcodes.js and update relevant imports
